+++ service/routes.py
@@ -5,7 +5,7 @@ Describe what your service does here
 """
 
 from flask import Flask, jsonify, request, url_for, make_response, abort
-from .utils import status  # HTTP Status Codes
+from .common import status  # HTTP Status Codes
 from service.models import YourResourceModel
 
 # Import Flask application
